Moves exceptions class into exceptions directory
This is because there will now be at least two exception related sets of functionality that lives here.

One for parsing exceptions.  And another for building out exceptions.

This move creates a sensible home for the new “build” related code to live.
